pyproject.toml
[project]
name = "moteus-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "moteus>=0.3.98",
    "opencv-python>=4.13.0.92",
    "pyyaml>=6.0.3",
    "ultralytics>=8.4.14",
]
